[newjersey/doula-pm#188] Don't require instructor training phone number
Instructor training phone number should not be required. However, previously an error message it would be displayed to the user if instructor training phone number was not in session storage.

This commit makes it okay for instructor training phone number to be null, and removes unnecessary type casts.

This commit resolves #[188](https://github.com/newjersey/doula-pm/issues/188).