[newjersey/doula-pm#206] Prefactor for whether a user is a supported sole proprietor
In anticipation of screening page 2, we expect to have 5 questions that collectively determine whether a user is a supported sole proprietor. If they are, then a bunch of pdf fields should be filled.

The current code just stores whether the user is a sole proprietor. Screening pages 2 and 3 expect to add these 4 additional questions that determine whether a user is "supported".

This commit refactors the `natureOfDisclosingEntity` enum property on `FormData`, to instead be a boolean `isSupportedSoleProprietor` property.

This commit contributes to #[206](https://github.com/newjersey/doula-pm/issues/206).