[newjersey/doula-pm#206] Screening page 2 part 2 - fill in pdf fields when the user is a supported sole
When a user answers all questions in the screening pages in a way that indicates that we support their use case, the ffs_individual.pdf should be filled according to the "Output to PDF" specifications described in https://github.com/newjersey/doula-pm/issues/175.

So far we have implemented the first two of three screening pages, http://localhost:3000/form/screening/1 and http://localhost:3000/form/screening/2. Previously, of the required fields to fill we had only implemented checking the Sole Proprietorship checkbox on page 16.

This commit
- Updates to fill in all pdf fields specified under "Output to PDF" in https://github.com/newjersey/doula-pm/issues/175, if a user answers the screening questions in a way that indicates that their use case is supported
-  In doing so, it fully types all possible pdf field names for all pages involved

It also
- Pulls out and reorganizes tests for page 16 of filling the pdf from `fillFfsIndividual.test.ts` to a separate file, fitting the pattern established in https://github.com/newjersey/doula-medicaid/pull/115
- Adds an assertion on the thrown error type to page 3 tests

This commit resolves #[206](https://github.com/newjersey/doula-pm/issues/206).