[newjersey/doula-pm#221] Create and use reusable address component
This commit creates a reusable address component, and uses it in personal details 2 and training.
It does not use this component in business details 1. That will be handled as part 2 of https://github.com/newjersey/doula-medicaid/pull/137.

This commit contributes to #[221](https://github.com/newjersey/doula-pm/issues/221) and #[180](https://github.com/newjersey/doula-pm/issues/180).

# Conflicts:
#	src/app/form/(formSteps)/personal-details/2/page.tsx