[newjersey/doula-pm#261] Rename all naming usages of session storage to data store.
This commit renames all naming usages of session storage to data store. It does not change where data is stored; data is still stored in session storage.

"data store" was chosen as a name instead of "store" because "store" can sound like a Redux store, and that's not what we're going for. Data store also seems generic enough to cover e.g. if we switch to a back end.

This commit is a prefactor to #[261](https://github.com/newjersey/doula-pm/issues/261).