[newjersey/doula-pm#290] Deploy on ECS part 1
Out of scope for this PR:
- Stable VPC-internal URL for load balancer, or other IaC configuration such that a stable URL is not needed
- Deploy on prod
- Readme improvements for docker container build tag and push, VPC validate, cdk deployment, and obtaining cloudformation artifacts

This commit adds a cdk stack to deploy the already-hosted application image to AWS ECS, within the hardcoded Dev environment VPC.

It also adds a script to configure the VPC, since the enableDnsSupport and enableDnsHostnames can only be set by cdk on resource creation.

This commit contributes to #[290](https://github.com/newjersey/doula-pm/issues/290).

Co-authored-by: John Chan <jachan@users.noreply.github.com>
Co-authored-by: Spencer Wohlers <seidior@users.noreply.github.com>