[newjersey/doula-pm#273] Confirm before leaving form
This commit adds a confirmation dialogue if the user has form data saved, and they try to refresh, close the tab, or navigate to a different website within the same tab.

The dialogue does not trigger if the user does not have form data saved.

This commit also refactors spies on `sendGAEvent` to be defined within relevant tests, instead of top-level in their test files, to more easily understand where the spies come from.

Co-authored-by: John Chan <jachan@users.noreply.github.com>

This commit resolves #[273](https://github.com/newjersey/doula-pm/issues/273).