[newjersey/doula-pm#292] Create action to manually trigger prod deployment
This commit creates the github action that can manually trigger production deployment.

Unfortunately, per https://docs.github.com/en/actions/reference/workflows-and-actions/events-that-trigger-workflows#workflow_dispatch :

> This event will only trigger a workflow run if the workflow file exists on the default branch.

So we will need to merge this before we can find out if it works. But it should be able to be triggered on any branch, as long as it exists in main.

This commit contributes to #[292](https://github.com/newjersey/doula-pm/issues/292).