[newjersey/doula-pm#329] Fix address bug
This commit fixes a bug where if address line 2 is not filled, then the city, state, and zip is also not filled.

It was caused because the code expected addressLine2 to be undefined if not filled. However, it is actually an empty string when not filled.

This commit also renames our address-formatting functions for clarity

This commit resolves #[329](https://github.com/newjersey/doula-pm/issues/329).