[newjersey/doula-pm#315] Enable end to end testing with feature flags on or off
Previously, we had set up our application deployment to load environment variables as feature flags, in anticipation of features that we will put behind flags.

Jest unit tests can mock `process.env.THE_FEATURE_FLAG`, or a getter function, to test both versions with and without. However, I don't believe this is possible with cypress - the cypress tests run on a different process than the application process. Yet, e2e testing both with feature flags on and off its important because we are likely to develop with all feature flags on, but release with a smaller set of feature flags on.

This commit enables running e2e testing on two circumstances: feature flags being on, or feature flags being off. For simplicity's sake this is just implementing that they are all on or off, but in reality the "on" state mimic development, and the "off" state should mimic the set of feature flags that we intend to release on production.

This commit does so by:
- Updating github actions to run the application server with these two sets off environment variables (`.env.test` with flags on, or `.env.test-flags-off`), and run the corresponding set of cypress tests against them (`npm run cypress:run`, or `npm run cypress:run:flagsOff`)
- Adding documentation about how to locally run these two sets of flags with their corresponding sets of cypress tests

Note that cypress tests which are not run are marked as "pending" rather than "skipped". I think it will be nicer if they were marked as "skipped", but it's not a big deal.

This commit resolves #[315](https://github.com/newjersey/doula-pm/issues/315).

# Conflicts:
#	.github/workflows/end_to_end_test.yml