[newjersey/doula-pm#66] Automatically post on a linked issue when an amplify branch is deployed on a PR
When we deploy a branch with `amplify` in a branch name, it automatically spins up an amplify deployment. This allows PMs to preview the PR before it is merged into main. However, the amplify URL is annoying to assemble. We need to get the branch name, the amplify app id, and then `/humanservices/dmahs/info/doulahelp`.

This repository is public. Although the amplify deployment is protected by amplify's auth, it's still not ideal to share the amplify deployment URL on this public repo.

This commit adds an automation to post on a PR's linked issue (which is on a private repo), with the amplify deployment URL when one is deployed.

This commit resolves newjersey/doula-pm#66.