[newjersey/doula-pm#347] Add example legal test
This commit enables testing with `NEXT_PUBLIC_FLAG_LEGAL` set to `"1"`, and includes an example legal test to stub this. It does so by changing the `getProgressBarSections` and `allSections` variables that define what sections we have, from being variables that are evaluated at file-read time, to functions that are evaluated dynamically.

This commit resolves newjersey/doula-pm#347.