Partially revert "[newjersey/doula-pm#66] Automatically post on a linked issue when an amplify branch is deployed on a PR"
This partially reverts commit 37bdf66b05b80d21c16bdbcd747ad68c8d52b220.

This commit also removes the automation to comment with the amplify build url. The ideas is nice, but it's too complicated in practice. E.g. how to not comment on the issue twice if a PR is drafted and re-opened.