[newjersey/doula-pm#335] Add Legal step 2
This commit adds the questions in Step 2 of the legal section, per the [ticket](https://github.com/newjersey/doula-pm/issues/335) and the [figma](https://www.figma.com/design/2R3VxYvZQYsaIxBcWoA1kL/Doula-Prototypes?node-id=5120-17655&t=jr8ZjLIjgxRKby1r-4).

It fails to make the pdf-writing conditional, causing the application to fail in the production application scenario when the legal flag is off.

This commit contributes to newjersey/doula-pm#335.