[newjersey/doula-pm#336] Add legal step 3
This commit adds the questions in Step 3, per the [ticket](https://github.com/newjersey/doula-pm/issues/335) and the [figma](https://www.figma.com/design/2R3VxYvZQYsaIxBcWoA1kL/Doula-Prototypes?node-id=5120-17655&t=jr8ZjLIjgxRKby1r-4). It writes the use responses to fields in page 9 of the pdf. All of this is conditional depending on if `process.env.NEXT_PUBLIC_FLAG_LEGAL` is set to `"1"`.

This commit also adds an additional test, testing that focus is correctly set on the first input with an error, even though that input might not be present on first page render.

This commit resolves newjersey/doula-pm#336.