[newjersey/doula-pm#226] Refactor error focus interface
react-hook-form has built-in functionality that keeps track of the order of inputs on the page, as determined by the order in which react-hook-form's `register` function is called on the component. It then has built-in functionality that sets focus on the first input with an error, if there is an error on submission.

Previously, we have overriden this behavior via the `mayHaveThreeOrMoreErrors` prop on `DoulaForm`. The prop was named as such, because it is otherwise not intuitive for a developer that potentially having three or more simultaneous errors -> need to show error summary -> need to manually set focus. The flip side is that the naming was not intuitive for what `mayHaveThreeOrMoreErrors` would control.

Additionally, we began using this prop in pages that did not have three or more errors, but where we needed to manually the focus order because react-hook-form would infer the order incorrectly. Specifically, when an input is conditionally rendered in between other inputs on the page, react-hook-form infers it as coming _after_ the other inputs, since `register` is only called on that input after it appears, despite it appearing between other inputs.

This commit thus refactors the `DoulaFormProps` interface to hopefully make these meanings and implications clearer. It:

- Renames `s/mayHaveThreeOrMoreErrors/showErrorSummary`
- Renames `s/orderedInputNames/manualFocusOrder`
- Enables `manualFocusOrder` to be set even when `showErrorSummary` is false, for situations where there may not be an error summary, but we want to manually set focus order due to conditionally rendered inputs
- It still enforces that `manualFocusOrder` must be set when `showErrorSummary` is true
- The three or more errors -> show error summary reasoning is no longer indicated in code

This commit contributes to newjersey/doula-pm#226.