[newjersey/doula-pm#413] Production stable DNS
This changes updates our cdk stack to use a stable DNS for production, in addition to the self-signed certificate for the random cloudformation-generated DNS. The reasoning for this is explained in the updated `deployment.md` docs.

It also updates the readme to include a link to the live site.

This commit contributes to newjersey/doula-pm#413.