[newjersey/doula-pm#332] Make feature flags available during nextjs build
Previously, no `.env` file was made available during the github actions -> docker build -> `npm run build`. The `.env` files were just put in the task definition and made available as shell-level environment variables. This worked for getting the feature flags in the `.env` file to show up at the `/api/health` endpoint.

However, it seems like nextjs does some sort of bundling/optimization, such that `NEXT_PUBLIC_FLAG_LEGAL` wouldn't actually worked to show our features unless it was also available during the docker build.

This commit adds permissions on the github actions IAM role to be able to pull the `configuration/.env` file. It also updates the deploy workflow to pull this file.

It also improves deployment documentation, though it does not mention this. Hoping to be able to take this out when we [switch to vite](https://github.com/newjersey/doula-pm/issues/428). Mentioned this in the ticket.

This commit resolves newjersey/doula-pm#332.