[newjersey/doula-pm#342] Pull out reusable address state and zip code components
This commit pulls out reusable components for address state and zip code. This is done in anticipation of personal details 4, which requires state and zip code inputs but not a full address with street 1 and street 2. The feature does also require a City address input, but City is straightforward enough that there's not much to encapsulate.

This commit resolves newjersey/doula-pm#342.