[newjersey/doula-pm#342] Rename `getDefaultBoolean` to `getBooleanString`
This commit renames the `s/getDefaultBoolean/getBooleanString` function. The function returned `"" | "true" | "false"`, for use in radio buttons, which require a string value even if there are only two radio buttons corresponding to Yes/No.

This rename is done in anticipation of using the `getBoolean` function that actually returns a boolean type, for initializing the default value of a checkbox in Personal Step 4.

This commit contributes to newjersey/doula-pm#342.